Reorganized project structure: Moved images to 'assets/images' folder, React components to 'src/components' folder and separated backend and frontend servers
