+++ PetLife/frontend/src/assets/images/icon.png
