+++ PetLife/frontend/src/components/Adopt/AdoptHeader/AdoptHeader.jsx
@@ -1,9 +1,9 @@
-import img1 from '../../assets/adopt-catimg.jpeg'
-import img2 from '../../assets/adopt-chikenimg.jpeg'
-import img3 from '../../assets/adopt-rabbitimg.jpeg'
-import img4 from '../../assets/adopt-cowimg.jpeg'
-import img5 from '../../assets/adopt-mouseimg.jpeg'
-import img6 from '../../assets/adopt-dogimg.jpeg'
+import img1 from '../../../assets/images/adopt-catimg.jpeg'
+import img2 from '../../../assets/images/adopt-chikenimg.jpeg'
+import img3 from '../../../assets/images/adopt-rabbitimg.jpeg'
+import img4 from '../../../assets/images/adopt-cowimg.jpeg'
+import img5 from '../../../assets/images/adopt-mouseimg.jpeg'
+import img6 from '../../../assets/images/adopt-dogimg.jpeg'
 import './AdoptHeader.css'
 
 import { useState, useEffect } from 'react'
