+++ PetLife/frontend/src/components/BoilerPlate/Footer/Footer.jsx
