+++ PetLife/frontend/src/components/Help/HelpHead/HelpHead.css
